requirements.txt
streamlit
pandas
numpy
seaborn
matplotlib
scipy
scikit-learn
statsmodelsgit